Rename blog related component directories
